Multiline text

+++ generatorService/generator/helpers.go
@@ -28,13 +28,11 @@ func LoadFont() (font.Face, *truetype.Font, error) {
 }
 
 func DrawText(c *freetype.Context, text string, x int, y int, face font.Face) {
-
-	iwidthf := 0
+	iwidthf := 0.0
 	for _, x := range text {
 		awidth, _ := face.GlyphAdvance(rune(x))
-		iwidthf += int(float64(awidth) / 64)
+		iwidthf += float64(awidth) / 64
 	}
-	// fmt.Printf("%+v\n", iwidthf)
 	pt := freetype.Pt(x-int(iwidthf), y)
 	c.DrawString(strings.ToUpper(text), pt)
 }
